% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLE4SCR.R
\name{PMLE4SCR}
\alias{PMLE4SCR}
\title{Two-stage pseudo maximum likelihood estimation (PMLE) for the copula-based analysis of semi-competing risks (SCR) data}
\usage{
PMLE4SCR(
  data,
  time,
  death,
  status_time,
  status_death,
  T.fmla = ~1,
  D.fmla = ~1,
  Gfun = list(T = "PH", D = "PH"),
  copula.family,
  copula.control = list(link = NULL, formula = ~1),
  initial
)
}
\arguments{
\item{data}{a data frame containing variables names in the formulas: \code{T.fmla}, \code{D.fmla}, and \code{formula} in \code{copula.control}.}

\item{time}{a character string specifying the variable name in the \code{data} for the time to a non-terminal event.}

\item{death}{a character string specifying the variable name in the \code{data} for the time to death.}

\item{status_time}{a character string specifying the variable name in the \code{data} for the censoring indicator of the non-terminal event: \code{1} indicates the non-terminal time is observed, and \code{0} indicates censored.}

\item{status_death}{a character string specifying the variable name in the \code{data} for the censoring indicator of death: \code{1} indicates the death time is observed, and \code{0} indicates censored.}

\item{T.fmla}{an object of class \code{\link[stats]{formula}}: a symbolic description of the regression model to be fitted for the marginal distribution of the non-terminal event time.}

\item{D.fmla}{an object of class \code{\link[stats]{formula}}: a symbolic description of the regression model to be fitted for the marginal distribution of the death time.}

\item{Gfun}{a list of two components \code{T} and \code{D}, both character strings specifying the link function in the SPT model for the non-terminal event and death, respectively: "PH" (proportional hazards as the default value) or "PO" (proportional odds).}

\item{copula.family}{a character string: "Clayton", "Gumbel", or "Frank", specifying the copula family to be used for the dependence between the bivariate event times.}

\item{copula.control}{a list of two components: \code{link} (a character string: "identity", "log", or "log-1", specifying the link function for the regression model of the copula parameter; if \code{link = NULL} (default), the link function will be the default function for the specified copula family: "log" for Clayton, "identity" for Frank, and "log-1" for Gumbel), and \code{formula} (an object of class \code{\link[stats]{formula}}: a symbolic description of the regression model to be fitted for the copula parameter under the specified link function; if \code{formula = ~ 1} (default), the copula parameter is a constant)}

\item{initial}{a numerical value or a vector of numerical values for the initial values of the copula parameter or the regression coefficients for the copula parameter.}
}
\value{
a list of the following components:
\itemize{
  \item{\code{gamma}: }{a data frame containing the estimate and robust standard error (SE) of the copula parameter or regression coefficients for the copula parameter.}
 \item{\code{gamma.cov}: }{a matrix containing the variance or variance-covariance matrix of the copula parameter or regression coefficients for the copula parameter.}
 \item{\code{betaT}: }{a data frame containing the estimate and robust SE of the regression coefficients for the marginal distribution of the non-terminal event time.}
 \item{\code{dLambdaT}: }{a data frame containing the estimate and robust SE of the jump size of the baseline function for the marginal distribution of the non-terminal event time.}
 \item{\code{thetaT.cov}: }{a matrix containing the variance-covariance matrix of the parameters for the marginal distribution of the non-terminal event time.}
 \item{\code{betaD}: }{a data frame containing the estimate and robust SE of the regression coefficients for the marginal distribution of the death time.}
 \item{\code{dLambdaD}: }{a data frame containing the estimate and robust SE of the jump size of the baseline function for the marginal distribution of the death time.}
 \item{\code{thetaD.cov}: }{a matrix containing the variance-covariance matrix of the parameters for the marginal distribution of the death time.}
 \item{\code{naive}}{a list containing naive estimates of \code{betaT}, \code{dLambdaT}, and \code{gamma} by estimating the marginal distribution of the non-terminal event time without the information on the between-event dependence.}
 \item{\code{call}: }{a list containing the specified values of input arguments \code{time}, \code{death}, \code{status_time}, \code{status_death}, \code{T.fmla}, \code{D.fmla}, \code{copula.family}, and the following two components:}
   \itemize{
     \item{\code{copula.link}: }{a list containing three R functions: "h.fun" (the link function used for the copula parameter), "dot.h.fun" (the first-order derivative of "h.fun"), and "ddot.h.fun" (the second-order derivative of "h.fun").}
     \item{\code{copula.fmla}: }{the specified value of \code{fomula} of the input argument \code{copula.control}.}
   }
 \item{\code{Par2Tau}: }{a list containing two R functions: "tau.alpha" (transformation from the copula parameter to Kendall's tau), and "Dtau.alpha" (the first-order derivative of "tau.alpha" function)}
}
}
\description{
Two-stage pseudo maximum likelihood estimation (PMLE) for the copula-based analysis of semi-competing risks (SCR) data
}
\examples{
\code{data(BMT, package = "SemiCompRisks")}
\code{data = BMT \%>\%
        mutate(g = factor(g, levels = c(2, 3, 1),
                    labels = c("AML-low", "AML-high", "ALL")))}
\code{myfit = PMLE4SCR(data, time = "T2", death = "T1",
                        status_time = "delta2", status_death = "delta1",
                        T.fmla = ~ g, D.fmla = ~ g,
                        copula.family = "Clayton",
                        copula.control = list(link = "identity", formula = ~ g),
                        initial = c(2, 0, 0))}
\code{myfit$gamma}
\code{myfit$betaT}

}
