% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSPT.R
\name{predict.fitSPT}
\alias{predict.fitSPT}
\title{Survival estimation under semi-parametric transformation regression model (SPT) for a given data}
\usage{
\method{predict}{fitSPT}(obj, newdata)
}
\arguments{
\item{obj}{an object of \code{\link[PMLE4SCR]{fitSPT}}.}

\item{newdata}{a data frame containing the same variables in the \code{formula} of \code{\link[PMLE4SCR]{fitSPT}}.}
}
\value{
a list of the following components:
\describe{
 \item{\code{surv}}{a data frame containing the estimate, model-based standard eror (SE), and robust SE for the survival probabilities for a given data.}
 \item{\code{Psi.surv}}{a data frame containing the estimated asymptotic expansion of the estimator of the survival probabilities.}
}
}
\description{
Survival estimation under semi-parametric transformation regression model (SPT) for a given data
}
